+++ src/routes/asset/AssetConnectPage.tsx
@@ -3,7 +3,7 @@ import { Link } from "react-router-dom";
 import MiniCircle from "../../assets/minicircle.svg?react";
 import Essential from "../../assets/required.svg?react";
 import Checked from "../../assets/checked.svg?react";
-import PurpleBtn from "../../components/PurpleBtn";
+import PurpleBtn from "../../components/common/PurpleBtn";
 
 const AssetConnectPage = () => {
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
