+++ src/routes/challenge/ChallengeDetailPage.tsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import Header from "../../components/Header";
+import Header from "../../components/common/Header";
 import EmotionSave from "../../assets/emotion-save.svg?react";
 import SmallLogo from "../../assets/small-profile.svg?react";
 import Date from "../../assets/date.svg?react";
@@ -7,7 +7,7 @@ import Check from "../../assets/check.svg?react";
 import Happy from "../../assets/happy-face.svg?react";
 import Sad from "../../assets/sad-face.svg?react";
 import Angry from "../../assets/angry-face.svg?react";
-import Navbar from "../../components/Navbar";
+import Navbar from "../../components/common/Navbar";
 
 const ChallengeDetailPage = () => {
   return (
