+++ src/routes/challenge/ChallengePage.tsx
@@ -3,7 +3,7 @@ import { useState } from "react";
 import Coin from "../../assets/coin30.svg?react";
 import User1 from "../../assets/user-icon.svg?react";
 import EmotionSave from "../../assets/emotion-save.svg?react";
-import Navbar from "../../components/Navbar";
+import Navbar from "../../components/common/Navbar";
 
 const ChallengePage = () => {
   const [category, setCategory] = useState("최신순");
